refactor(web): use proper dir name
