+++ apps/web/src/components/3d-scene/AnimatedBox.tsx
