+++ apps/web/src/components/3d-scene/Model.tsx
