+++ apps/web/src/components/3d-scene/SceneGradient.tsx
