+++ apps/web/src/components/3d-scene/SpeechBubble.tsx
