+++ apps/web/src/components/panels/AssetsPanel.tsx
